+++ CHANGELOG.md
@@ -16,6 +16,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Add LocalizedError conformance to SDKError and PackagerError for human-readable error descriptions
 
+### Changed
+
+- Packager, Request, ExecutionLog, and UploadCache accept FileSystemManager parameter
+
 ## 0.16.0 - 2026-03-30
 
 ### Removed